Redistributed workflows and consumers options and tests to their folders instead of grouped in core
