+++ src/packages/emmett/src/consumers/index.ts
@@ -1,1 +1,2 @@
 export * from './consumers';
+export * from './observability';
