+++ src/packages/emmett/src/consumers/observability/index.ts
@@ -1,0 +1,1 @@
+export * from './consumerCollector';
