--- src/packages/emmett/src/observability/collectors/index.ts
@@ -1,3 +1,0 @@
-export * from '../../eventStore/observability/eventStoreCollector';
-export * from './consumerCollector';
-export * from './workflowCollector';
