+++ src/packages/emmett/src/observability/index.ts
@@ -1,4 +1,3 @@
 export * from '@event-driven-io/almanac';
 export * from './attributes';
-export * from './collectors';
 export * from './options';
