+++ src/packages/emmett/src/workflows/index.ts
@@ -1,3 +1,4 @@
 export * from './handleWorkflow';
+export * from './observability';
 export * from './workflow';
 export * from './workflowProcessor';
